Update country-info.csv
Adding the lacking entry for the new University of Technology Nuremberg. Please note that UTN is the official short name of this university. Side note: "Technical University of Nuremberg" would be consistent with the naming of all other technical universities in Germany, but the short name TUN cannot be protected.